+++ src/js/ui/NewDataExploration/Layer/Constraints/AdvancedConstraints.js
@@ -57,37 +57,89 @@ You may add Your own copyright statement to Your modifications and may provide a
 END OF TERMS AND CONDITIONS
 */
 import React, {useState} from 'react';
-import FormGroup from '@material-ui/core/FormGroup';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import Checkbox from '@material-ui/core/Checkbox';
-import {componentIsRendering} from "../Sidebar";
-import Util from "../../library/apertureUtil"
+import { makeStyles } from '@material-ui/core/styles';
+import Modal from '@material-ui/core/Modal';
+import {Button, Paper} from "@material-ui/core";
+import SettingsIcon from "@material-ui/icons/Settings";
+import AdvancedConstraintCheckbox from "./AdvancedConstraintCheckbox";
+import {componentIsRendering} from "../../../Sidebar";
+import CloseIcon from '@material-ui/icons/Close';
+import Grid from "@material-ui/core/Grid";
 
-function updateLayerConstraints(activeLayerConstraints, index) {
-    let tempActiveConstraints = [...activeLayerConstraints];
-    tempActiveConstraints[index] = !tempActiveConstraints[index];
-    return tempActiveConstraints;
-}
+const useStyles = makeStyles((theme) => ({
+    modal: {
+        top: '20%',
+        left: '70%',
+        position: 'absolute',
+        width: 400,
+        backgroundColor: theme.palette.background.paper,
+        padding: theme.spacing(2),
+    },
+    constraintSection: {
+        overflowY: "auto",
+        maxHeight: "50vh",
+        padding: theme.spacing(1),
+    },
+    closeButtonSection: {
+        marginBottom: theme.spacing(1),
+        padding: theme.spacing(1),
+    },
+    closeButton : {
+        width: '100%',
+    },
+}));
 
-export default function AdvancedConstraintCheckbox(props) {
-    const [check, setCheck] = useState(props.activeLayerConstraints[props.constraintIndex]);
+export default function AdvancedConstraints(props) {
+    const classes = useStyles();
+    const [open, setOpen] = useState(false);
 
-    if(componentIsRendering) {console.log("|AdvancedContraintCheckbox Rerending|")}
+    if(componentIsRendering) {console.log("|AdvancedConstraints Rerending|")}
     return (
-        <FormGroup>
-            <FormControlLabel
-                control={
-                    <Checkbox
-                        checked={check}
-                        onChange={() => {
-                            setCheck(!check);
-                            props.setActiveLayerConstraints(updateLayerConstraints(props.activeLayerConstraints, props.constraintIndex));
-                        }}
-                        color="primary"
-                    />
-                }
-                label={props.constraint.label ?? Util.cleanUpString(props.constraint.name)}
-            />
-        </FormGroup>
+        <>
+            <Button variant="outlined" startIcon={<SettingsIcon/>} onClick={() => setOpen(true)}>
+                Advanced...
+            </Button>
+            <Modal
+                //FIXME According to https://material-ui.com/api/modal/ these should prevent the modal from 'focusing' but they don't
+                // disableEnforceFocus={true}
+                // disableAutoFocus={true}
+                aria-labelledby="adv-constraints"
+                open={open}
+                onClose={() => setOpen(false)}
+            >
+                <Grid
+                    id="adv-constraints"
+                    className={classes.modal}
+                    container
+                    direction="column"
+                    justifyContent="center"
+                    alignItems="stretch"
+                >
+                    <Grid item>
+                        <Paper className={classes.closeButtonSection} elevation={3}>
+                            <Button
+                                className={classes.closeButton}
+                                startIcon={<CloseIcon/>}
+                                variant="outlined"
+                                onClick={() => setOpen(false)}
+                            >
+                                Close
+                            </Button>
+                        </Paper>
+                    </Grid>
+                    <Grid item>
+                        <Paper elevation={3} className={classes.constraintSection}>
+                            {props.allLayerConstraints.map((constraint, index) => {
+                                return (
+                                    <div key={index}>
+                                        <AdvancedConstraintCheckbox activeLayerConstraints={props.activeLayerConstraints} setActiveLayerConstraints={props.setActiveLayerConstraints}
+                                                                    constraintIndex={index} constraint={constraint}/>
+                                    </div>)
+                            })}
+                        </Paper>
+                    </Grid>
+                </Grid>
+            </Modal>
+        </>
     );
 }
