+++ src/js/ui/NewDataExploration/Layer/Constraints/IndividualConstraint.js
