+++ src/js/ui/NewDataExploration/Layer/Layer.js
@@ -65,9 +65,9 @@ import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import {Grid, Paper, Switch} from "@material-ui/core";
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import LayerControls from "./LayerControls";
-import {componentIsRendering} from "../Sidebar";
-import AutoQuery from '../../library/autoQuery';
-import IndividualConstraint from "./IndividualConstraint"
+import {componentIsRendering} from "../../Sidebar";
+import AutoQuery from '../../../library/autoQuery';
+import IndividualConstraint from "./Constraints/IndividualConstraint"
 
 const useStyles = makeStyles((theme) => ({
     root: {
