+++ src/js/ui/NewDataExploration/Utils/Helpers.js
