+++ src/js/ui/NewDataExploration/WorkspaceControl/SaveLoadShare/Share.js
@@ -56,189 +56,92 @@ You may add Your own copyright statement to Your modifications and may provide a
 
 END OF TERMS AND CONDITIONS
 */
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import Typography from '@material-ui/core/Typography';
-import {componentIsRendering} from "../Sidebar";
+import {componentIsRendering} from "../../../Sidebar";
+import {
+    Switch,
+    FormGroup,
+    FormControlLabel,
+    Typography,
+    TextField,
+    Button,
+    Divider
+} from "@material-ui/core";
+import {CustomTooltip} from "../../../UtilityComponents";
 import Grid from "@material-ui/core/Grid";
-import { KeyboardDatePicker, KeyboardTimePicker } from '@material-ui/pickers';
-import { Alert } from "@material-ui/lab";
-import IconButton from "@material-ui/core/IconButton";
-import HighlightOffIcon from '@material-ui/icons/HighlightOff';
+import AssignmentTurnedInIcon from '@material-ui/icons/AssignmentTurnedIn';
+import CustomAlert from "../../Utils/CustomAlert";
 
-const configs = {
-    year: {
-        views: ["year"]
-    },
-    day: {
-        views: ["year", "month", "day"],
-        format: "MM/DD/yyyy"
-    },
-    "30min": {
-        views: ["year", "month", "day"],
-        format: "MM/DD/yyyy"
-    },
-}
-
-
-const useStyles = makeStyles({
-    root: {
-        width: '100%',
+const useStyles = makeStyles((theme) => ({
+    spaceOnTheLeft: {
+        marginLeft: theme.spacing(2),
     },
     title: {
-        textAlign: "center",
+        borderBottom: '2px solid #adadad',
+        marginBottom: theme.spacing(2),
+        width: "100%",
     },
-    nowrap: {
-        whiteSpace: "nowrap",
+    fullWidth: {
+        width: "100%",
     },
-    halfSize: {
-        width: '50%',
-    },
-});
+}));
 
-const epochToDate = (epoch) => {
-    const refrenceTime = new Date(epoch);
-    return new Date(epoch + refrenceTime.getTimezoneOffset() * 60000);
-}
+const aperturePublicPage = "urban-sustain.org/services/aperture.php"
 
-const dateToEpoch = (date) => {
-    return date.valueOf() - date.getTimezoneOffset() * 60000;
-}
-
-export default function ConstraintDate({ constraint, querier }) {
+export default React.memo(function Share({ serializeWorkspace, setModalOpen }) {
     const classes = useStyles();
-    const min = constraint.range[0];
-    const max = constraint.range[1];
-    const [minMaxDate, setMinMaxDate] = useState([epochToDate(min), epochToDate(max)]);
-    const config = configs[constraint.step];
+    const [includeColor, setIncludeColor] = useState(true)
+    const [includeViewport, setIncludeViewport] = useState(false)
+    const [linkCopied, setLinkCopied] = useState(false);
+    const [alertOpen, setAlertOpen] = useState(false);
+    const link = `${aperturePublicPage}?workspace=${serializeWorkspace("Shared Workspace", includeColor, includeViewport)}`
 
-    let errorMessageDisplay = [{ display: "none" }, { display: "block", width: '100%' }];
-    const [errorMessage, setErrorMessage] = useState(errorMessageDisplay[0]);
-    const [errorMessageText, setErrorMessageText] = useState("");
-
-    useEffect(() => {
-        const minMaxCommited = [dateToEpoch(minMaxDate[0]), dateToEpoch(minMaxDate[1])]
-        if (!constraint.forceUpdateFlag) {
-            constraint.state = minMaxCommited;
-        }
-        querier.updateConstraint(constraint.name, minMaxCommited);
-    }, [minMaxDate]);
-
-    useEffect(() => {
-        querier.constraintSetActive(constraint.name, true);
-        return () => {
-            querier.constraintSetActive(constraint.name, false);
-        }
-    }, []);
-
-    useEffect(() => {
-        if (constraint.forceUpdateFlag) {
-            constraint.forceUpdateFlag = false;
-            const newDate = [epochToDate(constraint.state[0]), epochToDate(constraint.state[1])]
-            setMinMaxDate(newDate);
-        }
-    })
-
-    function handleDateUpdate(e, setMin) {
-        if (e) {
-            const newTime = new Date(e.valueOf());
-            if (setMin && newTime > minMaxDate[1] || !setMin && newTime < minMaxDate[0]) {
-                setErrorMessageText(setMin ? "Min Date must be less than Max Date" : "Max Date must be greater than Min Date");
-                setErrorMessage(errorMessageDisplay[1]);
-            }
-            else {
-                setErrorMessage(errorMessageDisplay[0]);
-                setMin ? setMinMaxDate([new Date(e.valueOf()), minMaxDate[1]]) :
-                    setMinMaxDate([minMaxDate[0], new Date(e.valueOf())]);
-            }
-        }
+    const copyLink = () => {
+        const linkElement = document.getElementById("linkField");
+        linkElement.focus();
+        linkElement.select();
+        document.execCommand('copy');
+        setLinkCopied(true);
+        setAlertOpen(true);
     }
 
-    if (componentIsRendering) { console.log("|ContraintSlider Rerending|") }
+    if (componentIsRendering) { console.log("|Share Rerending|") }
     return (
-        <div className={classes.root} id={`constraint-div-${constraint.label}`}>
-            <Grid container direction="row" justifyContent="center" alignItems="center">
-                <Grid item>
-                    <Typography className={classes.title} id={`date-picker-${constraint.label}`} gutterBottom>
-                        <span className={classes.nowrap}>{minMaxDate[0].toDateString()} - {minMaxDate[1].toDateString()}</span>
-                    </Typography>
-                </Grid>
-            </Grid>
-            <Grid container direction="row" justifyContent="center" alignItems="center">
-                <Grid item className={classes.halfSize}>
-                    <KeyboardDatePicker
-                        label="Min Date"
-                        format="MM/DD/yyyy"
-                        value={minMaxDate[0]}
-                        minDate={epochToDate(min)}
-                        maxDate={minMaxDate[1]}
-                        onChange={(e) => {
-                            handleDateUpdate(e, true)
-                        }}
+        <Grid
+            container
+            direction="column"
+            justifyContent="center"
+            alignItems="flex-start"
+        >
+            <Grid item className={classes.fullWidth}>
+                <Typography className={classes.title} align="center" variant="h5">Share Workspace</Typography>
+                <Typography>Save Options</Typography>
+                <FormGroup row>
+                    <FormControlLabel
+                        control={<Switch className={classes.spaceOnTheLeft} color="primary" checked={includeColor} onChange={(e) => { setIncludeColor(e.target.checked) }} />}
+                        label="Include Color Selections"
                     />
-                </Grid>
-                <Grid item className={classes.halfSize}>
-                    <KeyboardDatePicker
-                        label="Max Date"
-                        format="MM/DD/yyyy"
-                        value={minMaxDate[1]}
-                        minDate={minMaxDate[0]}
-                        maxDate={epochToDate(max)}
-                        onChange={(e) => {
-                            handleDateUpdate(e, false)
-                        }}
+                </FormGroup>
+                <FormGroup row>
+                    <FormControlLabel
+                        control={<Switch className={classes.spaceOnTheLeft} color="primary" checked={includeViewport} onChange={(e) => { setIncludeViewport(e.target.checked) }} />}
+                        label="Include Current Viewport Bounds"
                     />
-                </Grid>
+                </FormGroup>
             </Grid>
-            {renderTime()}
-        </div>
+            <Grid item className={classes.fullWidth}>
+                <CustomTooltip title="Copy Link to Clipboard">
+                    <Button className={classes.fullWidth} variant="outlined" startIcon={<AssignmentTurnedInIcon/>} onClick={copyLink}>
+                        <Divider orientation="vertical" flexItem />
+                        &nbsp;&nbsp;
+                        <TextField className={classes.fullWidth} value={link} InputProps={{
+                            readOnly: true,
+                        }} id="linkField" />
+                    </Button>
+                </CustomTooltip>
+            </Grid>
+            <CustomAlert alertOpen={alertOpen} setAlertOpen={setAlertOpen} severity="success" text="Link Copied!" />
+        </Grid>
     );
-
-    function renderTime() {
-        if (constraint.step === "30min") {
-            return (
-                <div>
-                    <br/>
-                    <Grid container direction="row" justifyContent="center" alignItems="center">
-                        <Grid item className={classes.halfSize}>
-                            <KeyboardTimePicker
-                                label="Min Time"
-                                value={minMaxDate[0]}
-                                minDate={epochToDate(min)}
-                                maxDate={minMaxDate[1]}
-                                onChange={(e) => {
-                                    handleDateUpdate(e, true)
-                                }}
-                            />
-                        </Grid>
-                        <Grid item className={classes.halfSize}>
-                            <KeyboardTimePicker
-                                label="Max Time"
-                                value={minMaxDate[1]}
-                                minDate={minMaxDate[0]}
-                                maxDate={epochToDate(max)}
-                                onChange={(e) => {
-                                    handleDateUpdate(e, false)
-                                }}
-                            />
-                        </Grid>
-                    </Grid>
-                    {renderErrorMessage()}
-                </div>
-            );
-        }
-    }
-
-    function renderErrorMessage() {
-        return <div style={errorMessage}>
-            <br />
-            <Alert severity="error">{errorMessageText}
-                <span>
-                    <IconButton onClick={() => { setErrorMessage(errorMessageDisplay[0]) }}>
-                        <HighlightOffIcon />
-                    </IconButton>
-                </span>
-            </Alert>
-        </div>
-    }
-}
+})
