+++ src/js/ui/NewDataExploration/WorkspaceControl/WorkspaceControls.js
@@ -56,55 +56,102 @@ You may add Your own copyright statement to Your modifications and may provide a
 
 END OF TERMS AND CONDITIONS
 */
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import {componentIsRendering} from "../Sidebar";
-import Modal from "@material-ui/core/Modal";
-import Save from "./Save";
-import Load from "./Load";
-import Share from "./Share";
-import {Paper} from "@material-ui/core";
+import {Button, ButtonGroup, Grid, Paper, Switch, Icon, Fab} from "@material-ui/core";
+import SaveIcon from '@material-ui/icons/Save';
+import FolderOpenIcon from '@material-ui/icons/FolderOpen';
+import ShareIcon from '@material-ui/icons/Share';
+import WorkspaceSearchbar from "./WorkspaceSearchbar";
+import {componentIsRendering} from "../../Sidebar";
+import Ven from "../../../../../images/ven.svg"
+import VenFilled from "../../../../../images/venFilled.svg"
+import SaveAndLoadAndShare from './SaveLoadShare/SaveAndLoadAndShare';
+import EqualizerIcon from "@material-ui/icons/Equalizer";
+import CloseIcon from "@material-ui/icons/Close";
+import { useGlobalState } from "../../global/GlobalState";
+
 
 const useStyles = makeStyles((theme) => ({
-    paper: {
-        maxHeight: "95vh",
-        overflowY: "scroll",
-        top: '2.5vh',
-        left: '40%',
-        position: 'absolute',
-        width: 400,
-        backgroundColor: theme.palette.background.paper,
+    root: {
+        width: '100%',
         padding: theme.spacing(2),
-    }
+        margin: theme.spacing(1),
+    },
+    buttons: {
+        marginBottom: theme.spacing(2),
+    },
+    customIcon: {
+        width: "18px",
+        height: "18px",
+        transform: "translate(0, -10px)"
+    },
 }));
 
-export default React.memo(function SaveAndLoad({ mode, modalOpen, setModalOpen, serializeWorkspace, deSerializeWorkspace }) {
+export default React.memo(function WorkspaceControls(props) {
+    const [globalState, setGlobalState] = useGlobalState();
     const classes = useStyles();
+    const venIcon = <Icon>
+        <img src={props.intersect ? VenFilled : Ven} className={classes.customIcon} />
+    </Icon>
+    const [saveAndLoadAndShareModalOpen, setSaveAndLoadAndShareModalOpen] = useState(false)
+    const [saveAndLoadAndShareMode, setSaveAndLoadAndShareMode] = useState(null)
 
-    if (componentIsRendering) { console.log("|SaveAndLoad Rerending|") }
+    function handleDrawerClose() {
+        setGlobalState({sidebarOpen: false})
+    }
+
+    function toggleCharting() {
+        setGlobalState({ chartingOpen: !globalState.chartingOpen });
+    }
+
+    if (componentIsRendering) { console.log("|WorkspaceControls Rerending|") }
     return (
-        <Modal
-            open={modalOpen}
-            onClose={() => { setModalOpen(false) }}
-        >
-            <Paper className={classes.paper}>
-                {
-                    (() => {
-                        if (mode === "save") {
-                            return <Save serializeWorkspace={serializeWorkspace} setModalOpen={setModalOpen}/>
-                        }
-                        else if (mode === "load") {
-                            return <Load deSerializeWorkspace={deSerializeWorkspace} setModalOpen={setModalOpen}></Load>
-                        }
-                        else if (mode === "share") {
-                            return <Share serializeWorkspace={serializeWorkspace} setModalOpen={setModalOpen} />
-                        }
-                        else {
-                            return <div>Error: Invalid mode {mode}</div>
-                        }
-                    })()
-                }
+        <>
+            <Paper className={classes.root} elevation={3}>
+                <Grid className={classes.buttons} container direction="row" justifyContent="space-between" alignItems="center">
+                    <Grid item>
+                        <Button variant="outlined" startIcon={<SaveIcon />} onClick={() => {
+                            setSaveAndLoadAndShareModalOpen(true);
+                            setSaveAndLoadAndShareMode("save");
+                        }}>Save</Button>
+                    </Grid>
+                    <Grid item>
+                        <Button variant="outlined" startIcon={<FolderOpenIcon />} onClick={() => {
+                            setSaveAndLoadAndShareModalOpen(true);
+                            setSaveAndLoadAndShareMode("load");
+                        }}>Load</Button>
+                    </Grid>
+                    <Grid item>
+                        <Button variant="outlined" startIcon={<ShareIcon />} onClick={() => {
+                            setSaveAndLoadAndShareModalOpen(true);
+                            setSaveAndLoadAndShareMode("share");
+                        }}>Share</Button>
+                    </Grid>
+                    <Grid item>
+                        <Button variant="outlined" startIcon={venIcon} onClick={() => {
+                            props.setIntersect(!props.intersect)
+                        }}>
+                            {props.intersect ? "Intersections: on" : "Intersections: off"}
+                        </Button>
+                    </Grid>
+                    <Grid item>
+                        <Button variant="outlined" startIcon={<EqualizerIcon/>} id="nav-graph-button" onClick={() => toggleCharting()}>Graph</Button>
+                    </Grid>
+                    <Grid item>
+                        <Button variant="outlined" startIcon={<CloseIcon/>} onClick={handleDrawerClose}>Close</Button>
+                    </Grid>
+                </Grid>
+                <WorkspaceSearchbar layers={props.layers} graphableLayers={props.graphableLayers} layerTitles={props.layerTitles}
+                    workspace={props.workspace} setWorkspace={props.setWorkspace} />
             </Paper>
-        </Modal>
-    );
-})
+            <SaveAndLoadAndShare
+                modalOpen={saveAndLoadAndShareModalOpen}
+                setModalOpen={setSaveAndLoadAndShareModalOpen}
+                mode={saveAndLoadAndShareMode}
+                serializeWorkspace={props.serializeWorkspace}
+                deSerializeWorkspace={props.deSerializeWorkspace}
+            />
+        </>
+    )
+});
