+++ src/js/ui/NewDataExploration/WorkspaceLayers.js
@@ -58,7 +58,7 @@ END OF TERMS AND CONDITIONS
 */
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import Layer from "./Layer";
+import Layer from "./Layer/Layer";
 import {componentIsRendering} from "../Sidebar";
 
 const useStyles = makeStyles((theme) => ({
